move Task screen back
