+++ BrightID/src/bootstrap.js
@@ -2,10 +2,7 @@
 import { bootstrapAndUpgrade } from './versions';
 import { resetOperations } from './actions';
 import { store } from './store';
-import {
-  checkTasks,
-  syncStoreTasks,
-} from './components/SideMenu/Tasks/TasksSlice';
+import { checkTasks, syncStoreTasks } from './components/Tasks/TasksSlice';
 
 // happens inside of the loading screen
