+++ BrightID/src/components/Tasks/TaskCard.js
