+++ BrightID/src/components/Tasks/TaskCardController.js
