+++ BrightID/src/components/Tasks/TaskState.js
