+++ BrightID/src/components/Tasks/TasksProgress.js
