+++ BrightID/src/utils/operations.js
@@ -4,7 +4,7 @@ import api from '@/api/brightId';
 import store from '@/store';
 import { removeOperation, resetOperations, addLinkedContext } from '@/actions';
 import fetchUserInfo from '@/actions/fetchUserInfo';
-import { checkTasks } from '../components/SideMenu/Tasks/TasksSlice';
+import { checkTasks } from '../components/Tasks/TasksSlice';
 
 const time_fudge = 2 * 60 * 1000; // trace operations for 2 minutes
